+++ actix-session/src/storage/cookie.rs
@@ -1,6 +1,6 @@
+use super::SessionKey;
 use crate::storage::interface::{LoadError, SaveError, SessionState, UpdateError};
 use crate::storage::SessionStore;
-use crate::SessionKey;
 use std::convert::TryInto;
 use time::Duration;