+++ actix-session/CHANGES.md
@@ -2,6 +2,11 @@
 
 ## 0.6.0-beta.1 (Unreleased) - 2021-xx-xx
 
+### Changed
+* Rename `UserSession` to `SessionExt` [#212]
+  
+[#212]: https://github.com/actix/actix-extras/pull/212
+
 ## 0.5.0-beta.5 - 2021-12-12
 * Update `actix-web` dependency to `4.0.0.beta-14`. [#209]
 * Remove `UserSession` implementation for `RequestHead`. [#209]